Added option to load configurations like credentials from a config file
One may setup some default configurations and/or credentials in
/etc/boshifier/config.yml (or a config file location that is exported to
an environment variable BOSHIFIER_CONFIG_FILE). This would help:
* Set default configurations to be loaded from config file and thus no
 need to add repeated configs in the deployment specific configs
* One can set all bosh or/and concourse credentials in the config file
 so that end user doesnot need to worry about the credentials, they just
need to mention the targets with appropriate bosh names to make them
deployed.
* Even targets (the entire target map with the stages in it) also to be
loaded as configs so end user just need to provide deployment
configuration and it will be pushed to all targets configured